Add route deletion
We've two cases: where the prefix tree needs pruning, in which
case we erase nodes (and rows) as we go up the prefix path; or
alternatively, if we've deleted a node that had siblings, or come
to some other point that a route is set on, we stop pruning.

In either case, any node in the graph that still exists after the
pruning that was on the path to the deleted route will have had its
generation bumped. The consequence of this is that the route update
will be detected by the cache, which may reload some part of the
prefix tree on its next query.