Make generation updates more obvious
There was a small bug that'd affect the caching layer where a child
node could be created, deleted, then recreated, and it'd get the same
initial generation number (1). This'd confuse the cache layer which
wouldn't request a new version of the node.

Instead, since we're bumping generations, let's use the top-level
generation (of the app root) all the way down. Versions are still
incremented - they now just jump to the update number where they were
most recently changed.

This has the added bonus that eyeballing output is much simpler to
understand, without impacting on the cachability of untouched nodes.